==> 01-analise-arquivos.ipynb <==
https://dados.gov.br/dataset/serie-historica-de-precos-de-combustiveis-por-revenda

https://www.gov.br/anp/pt-br/assuntos/precos-e-defesa-da-concorrencia/precos/precos-revenda-e-de-distribuicao-combustiveis/levantamento-de-precos-de-combustiveis

import pandas as pd

!ls entrada/

!bzcat entrada/ca-2022-01.csv.bz2 | head -5

df = pd.read_csv("entrada/ca-2021-02.csv.bz2", sep=";", decimal=",", nrows=10000, encoding="iso8859")
#df = pd.read_csv("entrada/ca-2017-02.csv.bz2", sep=";", decimal=",", nrows=10000)
#df = pd.read_csv("entrada/ca-2022-01.csv.bz2", sep=";", nrows=10000, decimal=",")
df.head(5)

df.describe(include='all')

for col in df.columns:
    print(df[[col]].value_counts(), "\n")

df.columns

df2 = pd.read_csv("entrada/ca-2022-01.csv.bz2", sep=";", nrows=10000, decimal=",",
                 usecols=['Regiao - Sigla', 'Estado - Sigla', 'Municipio', 'Bairro',
                          'Produto', 'Valor de Venda', 'Valor de Compra', 'Unidade de Medida',
                          'Data da Coleta', 'Bandeira', 'Revenda'],
                 parse_dates=['Data da Coleta'], dayfirst=True)
df2.head(5)

%%script echo
df2.rename({'Regiao - Sigla': 'Regiao', 'Estado - Sigla': 'UF', 
            'Valor de Venda': 'Venda', 'Valor de Compra': 'Compra',
            'Unidade de Medida': 'Unidade', 'Data da Coleta': 'Coleta'},
           axis=1, inplace=True)
df2.head(5)

%%script echo
ocols = df2.columns.tolist()
ocols

%%script echo
lcols = [col.lower() for col in ocols]
lcols

%%script echo
dict(zip(ocols, lcols))

df2.rename({'Regiao - Sigla': 'regiao', 'Estado - Sigla': 'uf',
            'Municipio': 'municipio', 'Bairro': 'bairro',
            'Produto': 'produto', 'Valor de Venda': 'venda', 
            'Valor de Compra': 'compra', 'Unidade de Medida': 'unidade',
            'Data da Coleta': 'coleta',
            'Bandeira': 'bandeira',
            'Revenda': 'revenda'},
           axis=1, inplace=True)
df2.head(5)

df2.columns

df3 = df2[['regiao', 'uf', 'municipio', 'bairro',
     'produto', 'venda', 'compra', 'unidade', 'coleta',
     'bandeira', 'revenda']]
df3.head()
