==> 02-preparo-dados.ipynb <==
!cd entrada; ls *.csv.bz2

!rm -rf saida && mkdir saida

import pandas as pd
import os

def processar_arquivo_csv(arquivo_entrada, diretorio_saida="saida"):

    arquivo_saida = os.path.join(diretorio_saida, os.path.basename(arquivo_entrada))
    print("%s -> %s" % (arquivo_entrada, arquivo_saida))
    
    # ler arquivo CSV de entrada
    df1 = pd.read_csv(arquivo_entrada, sep=";", decimal=",", encoding="iso8859",
                     #usecols=['Regiao - Sigla', 'Estado - Sigla', 'Municipio', 'Bairro',
                     #         'Produto', 'Valor de Venda', 'Valor de Compra', 'Unidade de Medida',
                     #         'Data da Coleta', 'Bandeira', 'Revenda'],
                     parse_dates=['Data da Coleta'], dayfirst=True,
                     #nrows=10000
                    )
    
    # renomear colunas originais
    df1.rename({'ï»¿Regiao - Sigla': 'regiao',
                'Regiao - Sigla': 'regiao', 'Estado - Sigla': 'uf',
                'Municipio': 'municipio', 'Bairro': 'bairro',
                'Produto': 'produto', 'Valor de Venda': 'venda', 
                'Valor de Compra': 'compra', 'Unidade de Medida': 'unidade',
                'Data da Coleta': 'coleta',
                'Bandeira': 'bandeira',
                'Revenda': 'revenda'},
              axis=1, inplace=True)
    
    # extrair campos individuais da data
    df1['ano'] = df1['coleta'].dt.year
    df1['mes'] = df1['coleta'].dt.month
    df1['dia'] = df1['coleta'].dt.day
    
    # mudar a ordem de apresentação das colunas
    df2 = df1[['regiao', 'uf', 'municipio', 'bairro',
               'produto', 'venda', 'compra', 'unidade',
               'ano', 'mes', 'dia',
               'bandeira', 'revenda']]
    
    # gravar CSV resultante
    df2.to_csv(arquivo_saida, index=False, encoding='utf-8')
    
    # liberar memória
    del(df1)
    del(df2)

#processar_arquivo_csv("entrada/ca-2017-02.csv.bz2")

#df2 = pd.read_csv("saida/ca-2017-02.csv.bz2")
#df2

#df1 = pd.read_csv("entrada/ca-2017-02.csv.bz2", sep=";", decimal=",", encoding="iso8859",
#                  parse_dates=['Data da Coleta'], dayfirst=True)
#df1.info()

import glob

for arquivo in glob.glob("entrada/*.csv.bz2"):
    processar_arquivo_csv(arquivo)

!ls -lah saida/

from pathlib import Path

def gerar_unico_arquivo_csv(arquivo_saida, diretorio_entrada="saida"):
    df = pd.DataFrame()
    
    for arquivo in Path(diretorio_entrada).rglob('*.csv.bz2'):
        
        print(arquivo.name)
        dft = pd.read_csv(arquivo)
        
        if not df.empty:
            df = pd.concat([df, dft], ignore_index=True)
        else:
            df = dft
            
    if not df.empty:
        df.to_csv(arquivo_saida, index=False)

gerar_unico_arquivo_csv("ca-2015-2022.csv.bz2")

!ls -lha *.csv.bz2

df = pd.read_csv("ca-2015-2022.csv.bz2", nrows=1000)
df
