==> 03-analise-exploratoria.ipynb <==
https://www.gov.br/anp/pt-br/assuntos/precos-e-defesa-da-concorrencia/precos/precos-revenda-e-de-distribuicao-combustiveis/levantamento-de-precos-de-combustiveis

https://dados.gov.br/dataset/serie-historica-de-precos-de-combustiveis-por-revenda

import pandas as pd

!bzcat ca-2015-2022.csv.bz2 | head -5

# ler os dados do arquivo CSV
df = pd.read_csv("ca-2015-2022.csv.bz2",
                 dtype={'regiao': 'category', 'uf': 'category', 
                        'municipio': 'category', 'bairro': 'category',
                        'produto': 'category', 'venda': 'float32',
                        'compra': 'float32', 'unidade': 'category',
                        'ano': 'category', 'mes': 'category', 'dia': 'category'})

df

df.info()

df.describe(include='all')

for col in df.columns:
    print(df[[col]].value_counts(), "\n")

total = len(df)
for col in df.columns:
    qtd = df[df[col].isna()]['uf'].count()
    print("%10s: %7d nulos (%6.2f%%)" % (col, qtd, qtd / total * 100))
print("%10s: %7d linhas" % ("total", total))

df.groupby(['produto', 'unidade'])[['venda']].count().T

df[~df['compra'].isna()].\
    groupby(['ano'])[['compra']].count().T

df[df['produto'].isin(['GASOLINA', 'ETANOL'])].\
    groupby(['ano'])[['compra']].count().T

> **O contrato atual Levantamento de Preços de  Combustíveis (LPC) não prevê a coleta dos preços de compra dos revendedores junto às distribuidoras.**

#df.groupby(['regiao', 'uf', 'municipio', 'bairro'])[['venda']].count()

pd.set_option('display.max_rows', None)

pd.set_option('display.max_rows', 50)
