==> 04-filtro-dados.ipynb <==
import pandas as pd

!bzcat ca-2015-2022.csv.bz2 | head -5

# ler os dados do arquivo CSV
df = pd.read_csv("ca-2015-2022.csv.bz2",
                 dtype={'regiao': 'category', 'uf': 'category', 
                        'municipio': 'category', 'bairro': 'category',
                        'produto': 'category', 'venda': 'float32',
                        'compra': 'float32', 'unidade': 'category',
                        'ano': 'uint16', 'mes': 'uint8', 'dia': 'uint8'})

# manter apenas 2 tipos de produtos e remover unidade de medida
df = df[df['produto'].isin(['GASOLINA', 'ETANOL'])].drop('unidade', axis=1)
df

# remover campo dia (irrelevante para a análise)
df.drop('dia', axis=1, inplace=True)
df

# manter apenas medições até o ano de 2020
df = df[df['ano'] <= 2020]
df

# remover campos valor de compra (73% nulos) e dia (irrelevante para a análise)
#df.drop(['compra', 'dia'], axis=1, inplace=True)
#df

df.info()

df.describe(include='all')

# gerar arquivo CSV filtrado
arquivo_saida = "ge-2015-2020.csv.bz2"
df.to_csv(arquivo_saida, index=False, encoding='utf-8')
