==> 05-imputacao-dados.ipynb <==
import pandas as pd

# ler os dados do arquivo CSV
df = pd.read_csv("ge-2015-2020.csv.bz2",
                 dtype={'regiao': 'category', 'uf': 'category', 
                        'municipio': 'category', 'bairro': 'category',
                        'produto': 'category', 'venda': 'float32',
                        'compra': 'float32', 'unidade': 'category',
                        'ano': 'uint16', 'mes': 'uint8',
                        'bandeira': 'category'})

df

# selecionar apenas linhas que possuem valor de compra preenchido
df2 = df[~df['compra'].isna()]
df2

# remover colunas que serão desprezadas
#df2r = df2.drop(['regiao', 'municipio', 'bairro', 'bandeira', 'revenda'], axis=1)
df2r = df2.drop(['uf', 'municipio', 'bairro', 'bandeira', 'revenda'], axis=1)
df2r

df2r.describe(include='all')

df2r.corr()

# converter colunas categóricas em numéricas (one-hot encoding)
df2r = pd.get_dummies(df2r)
df2r

df2r.info()

df2r.corr()

## Treinar modelo de regressão

from sklearn.ensemble import RandomForestRegressor

X = df2r.drop('compra', axis=1).values
y = df2r['compra']

X.shape, y.shape

modelo = RandomForestRegressor(random_state=42, n_jobs=-1)
modelo

%time modelo.fit(X, y)

X[0]

modelo.predict([X[0]])

## Imputar valores ausentes

# selecionar apenas linhas que não possuem valor de compra preenchido
df3 = df[df['compra'].isna()]
df3

# remover colunas que serão desprezadas
#df3r = df3.drop(['regiao', 'municipio', 'bairro', 'bandeira', 'revenda', 'compra'], axis=1)
df3r = df3.drop(['uf', 'municipio', 'bairro', 'bandeira', 'revenda', 'compra'], axis=1)
df3r

# converter colunas categóricas em numéricas (one-hot encoding)
df3r = pd.get_dummies(df3r)
df3r

df3r.info()

df3r.corr()

X3 = df3r.values
X3

y3_pred = modelo.predict(X3)
y3_pred

# imputar os valores ausentes
df3['compra'] = y3_pred
df3

df3.corr()

## Gerar dataframe consolidado

# fazer marcação de que certos valores foram imputados
df = df2.copy()
df['imputado'] = 0
df = pd.concat([df, df3])
df['imputado'] = df['imputado'].fillna(1)
df['imputado'] = df['imputado'].astype('uint8')
df

#df = pd.concat([df2, df3])
#df

# arredondar valores para 4 casas decimais
for col in ['venda', 'compra']:
    df[col] = df[col].round(3)
df

# gerar arquivo CSV com valores imputados
arquivo_saida = "ge-cv-2015-2020.csv.bz2"
df.to_csv(arquivo_saida, index=False, encoding='utf-8')
