==> 06-analise-exploratoria2.ipynb <==
import pandas as pd

from funcoes_apoio import preparar_dataframe

df = preparar_dataframe()
df

df.info()

df.describe(include='all')

for col in df.columns:
    print(df[[col]].value_counts(), "\n")

%%script true
df2 = df[['bandeira', 'regiao']].groupby('bandeira').count().\
    reset_index().rename({'regiao': 'qtde'}, axis=1).\
    sort_values('qtde', ascending=False)
df2['perc'] = round(df2['qtde'] / len(df2), 2)
df2

%%script true
def normalizar_bandeira(titulo):
    if titulo in ['BRANCA', 'IPIRANGA', 'RAIZEN']:
        return titulo
    elif titulo == 'PETROBRAS DISTRIBUIDORA S.A.':
        return 'PETROBRAS'
    else:
        return 'OUTRAS'

df2['bandeira'].apply(normalizar_bandeira).head()

df['data'] = pd.to_datetime(df['ams'], format='%Y%m')
df

pd.to_datetime("202009", format='%Y%m')
